requirements.txt
Django
Pillow
python-dateutil
flake8
flake8-isort
black
django-modeltranslation
django-ckeditor
django-rosetta
python-dotenv